requirements.txt
streamlit
openai
pillow
pytesseract
numpy
matplotlib
PyPDF2